Test_prom/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest~=4.0.0
requests==2.25.1
selenium==3.141.0
toml==0.10.2
urllib3==1.26.3
webdriver-manager==3.3.0

setuptools~=57.4.0
pyodbc~=4.0.32
paramiko~=2.7.2
seletools~=1.2.1
allure-pytest~=2.9.43
scipy~=1.7.1
pydub~=0.25.1
pywinauto~=0.6.8
PyAutoGUI~=0.9.52